MINOR: quic-be: modify ssl_sock_srv_try_reuse_sess() to reuse backend session (0-RTT)
This function is called for both TCP and QUIC connections to reuse SSL sessions
saved by ssl_sess_new_srv_cb() callback called upon new SSL session creation.

In addition to this, a QUIC SSL session must reused the ALPN and some 0-RTT QUIC
transport parameters. This is what is added by this patch for QUIC 0-RTT SSL sessions.

Note that for now on, ssl_sock_srv_try_reuse_sess() may fail for QUIC connection
if it did not managed to reuse the ALPN. The caller must be informed of such an
issue. It must not enable 0-RTT for the current session in this case. This is
impossible without ALPN which is required to start a mux.

ssl_sock_srv_try_reuse_sess() always succeeds for TCP connections.