MEDIUM: server: do not remove idle conns in delete server
Removes idle connections removal from "delete server" handler. This
concerns both idle, purgeable and private connections.

Thus it is not safe anymore to remove a server if there is idle
connections attached into it. To ensure this cannot be performed,
srv_check_for_deletion() has been extended with checks on server members
<curr_idle_conns> counter and <sess_conns> list.

The main objective of this patch is to reduce the amount of work
conducted under thread isolation. Also, it ensures that there is no
issue when removing a dynamic QUIC server as quic MUX does not currently
implement a takeover callback.

As a side-effect, it may be more difficult to remove a server after
utilisation due to idle connection presence. As such, a future patch
will implement idle connections removal in "shutdown sessions server"
handler.