MAJOR: server: do not remove idle conns in delete server
Removes idle connections removal from "delete server" handler. This
concerns both idle, purgeable and private connections.

Thus it is not safe anymore to remove a server if there is idle
connections attached into it. To ensure this cannot be performed,
srv_check_for_deletion() has been extended with checks on server members
<curr_idle_conns> counter and <sess_conns> list.

The main objective of this patch is to reduce the amount of work
performed under thread isolation. Now, idle connection purging must be
scheduled prior to "delete server" invokation, thus switching from a
synchronous to asynchronous process.

This patch is considered as major as it may have an impact on "delete
server" execution. It could be slightly longer to be able to remove a
server, after its idle connections are removed. However, a previous
change has introduced idle connections cleanup, both standard and
private, when a server is switched to maintenance. Also, it is not
possible for a server in maintenance to keep new idle connections. As
such, it is expected that the difference for "delete server" completion
should be minimal.

Also, another objective of this patch is to be able to support dynamic
servers with QUIC, as takeover is not yet implemented for this protocol.
However, there is still some changes to perform to protect IO handler
for QUIC MUX.